+++ ts/packages/dispatcher/src/context/dispatcher/handlers/requestCommandHandler.ts
@@ -11,7 +11,6 @@ import {
     FullAction,
     ProcessRequestActionResult,
     ExplanationOptions,
-    ParamObjectType,
     equalNormalizedParamObject,
 } from "agent-cache";
 
@@ -60,8 +59,8 @@ import {
 import { DispatcherName } from "../../interactiveIO.js";
 import { getActionTemplateEditConfig } from "../../../translation/actionTemplate.js";
 import { getActionSchema } from "../../../translation/actionSchemaFileCache.js";
-import { isUnknownAction } from "../dispatcherAgent.js";
 import { UnknownAction } from "../schema/dispatcherActionSchema.js";
+import { isUnknownAction } from "../dispatcherUtils.js";
 
 const debugTranslate = registerDebug("typeagent:translate");
 const debugSemanticSearch = registerDebug("typeagent:translate:semantic");
@@ -724,108 +723,6 @@ export async function translateRequest(
     return requestAction;
 }
 
-async function canExecute(
-    requestAction: RequestAction,
-    context: ActionContext<CommandHandlerContext>,
-): Promise<boolean> {
-    const actions = requestAction.actions;
-    const systemContext = context.sessionContext.agentContext;
-    const unknown: UnknownAction[] = [];
-    const disabled = new Set<string>();
-    for (const action of actions) {
-        if (isUnknownAction(action)) {
-            unknown.push(action);
-        }
-        if (
-            action.translatorName &&
-            !systemContext.agents.isActionActive(action.translatorName)
-        ) {
-            disabled.add(action.translatorName);
-        }
-    }
-
-    if (unknown.length > 0) {
-        const unknownRequests = unknown.map(
-            (action) => action.parameters.request,
-        );
-        const lines = [
-            `Unable to determine ${actions.action === undefined ? "one or more actions in" : "action for"} the request.`,
-            ...unknownRequests.map((s) => `- ${s}`),
-        ];
-        systemContext.chatHistory.addEntry(
-            lines.join("\n"),
-            [],
-            "assistant",
-            systemContext.requestId,
-        );
-
-        const config = systemContext.session.getConfig();
-        if (
-            !config.translation.switch.search &&
-            !config.translation.switch.embedding &&
-            !config.translation.switch.inline
-        ) {
-            lines.push("");
-            lines.push("Switching agents is disabled");
-        } else {
-            const entries = await Promise.all(
-                unknownRequests.map((request) =>
-                    systemContext.agents.semanticSearchActionSchema(
-                        request,
-                        1,
-                        () => true, // don't filter
-                    ),
-                ),
-            );
-            const schemaNames = new Set(
-                entries
-                    .filter((e) => e !== undefined)
-                    .map((e) => e![0].item.actionSchemaFile.schemaName)
-                    .filter(
-                        (schemaName) =>
-                            !systemContext.agents.isSchemaActive(schemaName),
-                    ),
-            );
-
-            if (schemaNames.size > 0) {
-                lines.push("");
-                lines.push(
-                    `Possible agent${schemaNames.size > 1 ? "s" : ""} to handle the request${unknownRequests.length > 1 ? "s" : ""} are not active: ${Array.from(schemaNames).join(", ")}`,
-                );
-            }
-        }
-
-        displayError(lines, context);
-        return false;
-    }
-
-    if (disabled.size > 0) {
-        const message = `Not executed. Action disabled for ${Array.from(disabled.values()).join(", ")}`;
-        systemContext.chatHistory.addEntry(
-            message,
-            [],
-            "assistant",
-            systemContext.requestId,
-        );
-
-        displayWarn(message, context);
-        return false;
-    }
-
-    return true;
-}
-
-async function requestExecute(
-    requestAction: RequestAction,
-    context: ActionContext<CommandHandlerContext>,
-) {
-    if (!(await canExecute(requestAction, context))) {
-        return;
-    }
-
-    await executeActions(requestAction.actions, context);
-}
-
 async function canTranslateWithoutContext(
     requestAction: RequestAction,
     usedTranslators: Map<string, TypeAgentTranslator<object>>,
@@ -1096,16 +993,15 @@ export class RequestCommandHandler implements CommandHandler {
                 .history
                 ? getChatHistoryForTranslation(systemContext)
                 : undefined;
-            if (history) {
-                // prefetch entities here
-                systemContext.chatHistory.addEntry(
-                    request,
-                    [],
-                    "user",
-                    systemContext.requestId,
-                    cachedAttachments,
-                );
-            }
+
+            // prefetch entities here
+            systemContext.chatHistory.addEntry(
+                request,
+                [],
+                "user",
+                systemContext.requestId,
+                cachedAttachments,
+            );
 
             // Make sure we clear any left over streaming context
             systemContext.streamingActionContext = undefined;
@@ -1144,7 +1040,7 @@ export class RequestCommandHandler implements CommandHandler {
                     timestamp: new Date(),
                 });
             }
-            await requestExecute(requestAction, context);
+            await executeActions(requestAction.actions, context);
             if (canUseCacheMatch) {
                 await requestExplain(
                     requestAction,
