+++ ts/packages/dispatcher/src/rpc/dispatcherTypes.ts
@@ -6,15 +6,15 @@ import {
     DynamicDisplay,
     TemplateSchema,
 } from "@typeagent/agent-sdk";
-import { RequestMetrics } from "../utils/metrics.js";
 import { CommandCompletionResult } from "../command/completion.js";
+import { CommandResult } from "../dispatcher.js";
 
 export type DispatcherInvokeFunctions = {
     processCommand(params: {
         command: string;
         requestId?: string | undefined;
         attachments?: string[] | undefined;
-    }): Promise<RequestMetrics | undefined>;
+    }): Promise<CommandResult | undefined>;
 
     getDynamicDisplay(params: {
         appAgentName: string;
